CalculationsService: Changed types to float

+++ virtual-threads-app/src/main/java/com/dzydowicz/virtualthreads/application/dto/calculations/CalculateKineticEnergyResponse.java
@@ -8,5 +8,5 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class CalculateKineticEnergyResponse {
-    private Double kineticEnergy;
+    private Float kineticEnergy;
 }
